refactor(api): move answer routes and controller
