+++ api/src/evaluation/routes.js
@@ -1,3 +1,4 @@
+import * as answersRoutes from './application/answers/index.js';
 import * as autonomousCoursesRoutes from './application/autonomous-courses/index.js';
 import * as competenceEvaluationsRoutes from './application/competence-evaluations/index.js';
 import * as feedbacksRoutes from './application/feedbacks/index.js';
@@ -7,6 +8,7 @@ import * as stagesRoutes from './application/stages/index.js';
 import * as stageCollectionRoutes from './application/stage-collections/index.js';
 
 const evaluationRoutes = [
+  answersRoutes,
   autonomousCoursesRoutes,
   competenceEvaluationsRoutes,
   stageCollectionRoutes,
