+++ api/tests/evaluation/acceptance/application/answers/answer-controller-find_test.js
@@ -1,5 +1,5 @@
-import { expect, generateValidRequestAuthorizationHeader, databaseBuilder } from '../../../test-helper.js';
-import { createServer } from '../../../../server.js';
+import { expect, generateValidRequestAuthorizationHeader, databaseBuilder } from '../../../../test-helper.js';
+import { createServer } from '../../../../../server.js';
 
 describe('Acceptance | Controller | answer-controller-find', function () {
   describe('GET /api/answers?challengeId=Y&assessmentId=Z', function () {
